Actualizando la estructura y cambio de firestore
