+++ .gitignore
@@ -31,7 +31,7 @@ migrate_working_dir/
 .pub-cache/
 .pub/
 /build/
-
+/lib/firebase_options.dart
 # Symbolication related
 app.*.symbols
