--- lib/firebase_options.dart
@@ -1,81 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      case TargetPlatform.windows:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for windows - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      default:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyD80oGIMcQsmkcYwOjgqG42i1aDuI7ethM',
-    appId: '1:1037522055491:web:008047f7d1e400d3753ff6',
-    messagingSenderId: '1037522055491',
-    projectId: 'univallemovil-f834a',
-    authDomain: 'univallemovil-f834a.firebaseapp.com',
-    storageBucket: 'univallemovil-f834a.appspot.com',
-    measurementId: 'G-BJJ72V080K',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyDMvS4xdhwCam-3rMxFPwETC9lxRooMCi8',
-    appId: '1:1037522055491:android:61d6905781dbfd68753ff6',
-    messagingSenderId: '1037522055491',
-    projectId: 'univallemovil-f834a',
-    storageBucket: 'univallemovil-f834a.appspot.com',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyB1GG0nZWXOgtikhkfW2q6lAV0f0RPaAuc',
-    appId: '1:1037522055491:ios:43d85652c5cd355b753ff6',
-    messagingSenderId: '1037522055491',
-    projectId: 'univallemovil-f834a',
-    storageBucket: 'univallemovil-f834a.appspot.com',
-    iosBundleId: 'com.firebase.fluttermovil.firebaseCrud',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyB1GG0nZWXOgtikhkfW2q6lAV0f0RPaAuc',
-    appId: '1:1037522055491:ios:b2353e3c6ef91bc0753ff6',
-    messagingSenderId: '1037522055491',
-    projectId: 'univallemovil-f834a',
-    storageBucket: 'univallemovil-f834a.appspot.com',
-    iosBundleId: 'com.firebase.fluttermovil.firebaseCrud.RunnerTests',
-  );
-}
