+++ lib/screens/AnimalRegistrationScreen.dart
