--- lib/widgets/widget.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MiWidget extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      // Personaliza este widget según tus necesidades
-    );
-  }
-}
